[WS2-F1] refactor: shared llm_json.extract_json; harden vocab response parsing
